Settings: Allow for pointers into the EEPROM buffer
Do not copy every string out of the EEPROM object into a separate buffer. Instead just create a buffer object with a pointer into the EEPROM handle buffer, if it is a read-only operation. To prevent messing up the EEPROM data and the checksum, the returned buffer object should be taken as `const auto`.
This change will come in handy for reading DER cert and key files stored in the EEPROM, which only need to be read by BearSSL during start up. In the future all of the EEPROM buffer memory should be freed by calling `end()` after startup.