NoStl: Add type erasing function reference class
The `FunctionRef<>` class allows taking lambdas that capture variables without the need to make them templates. The class uses an internal template specialized through the constructor parameter and is type erased using virtual methods. Compared to `std::function<>` no copy of the lambda is created. Instead only a non-owning reference is used to remove the need for dynamic allocation. This means the user needs to be aware of the referenced lambda's life-time/scope, similar to `Buffer` objects created through slicing.