WebPage: Implement symmetric encryption
As HTTPS is dysfunctional, a custom encryption scheme is introduced based on the Chacha20 Poly1305 cipher which is provided by the Arduino libraries thanks to BearSSL.
During the setup wizard a 16 byte key is set or automatically generated, that is later used to derive one time keys using HKDF with a one time salt. The web browser now has to execute a considerable amount of JS code for en-/decryption, as JS crypto functionality is not available in an insecure HTTP environment. Encryption nonce values double as form nonce values to prevent playback attacks.